+++ tests/testthat/test-arc_raster.R
@@ -1,5 +1,7 @@
 test_that("arc_raster() works", {
   skip_on_cran()
+
+  unset_arc_token()
   img_url <- "https://landsat2.arcgis.com/arcgis/rest/services/Landsat/MS/ImageServer"
 
   landsat <- arc_open(img_url)